requirements.txt
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0
openai>=1.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-google-genai>=1.0.0
